KYLIN: arm64/kasan: fix check_memory_region renamed
Mainline: KYLIN-only
Severity: Low

Task: 216207

check_memory_region has been rename as kasan_check_range, let's do it
for mm/kasan/shadow.c

K2CI-Arch: All
Fixes: 63be85ac275c ("OPENEULER: arm64: introduce copy_mc_to_kernel() implementation")
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/109175
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>

+++ arch/arm64/include/asm/uaccess.h
@@ -438,7 +438,7 @@ static inline size_t probe_subpage_writeable(const char __user *uaddr,
 static inline unsigned long __must_check
 copy_mc_to_kernel(void *to, const void *from, unsigned long size)
 {
-	int ret;
+	unsigned long ret;
 
 	ret = memcpy_mcs(to, from, size);
 	return (ret == -EFAULT) ? size : 0;
